Prepare some crates for release (#1504)

+++ bee-api/bee-rest-api/src/routes/api/core/v2/blocks_submit.rs
@@ -131,7 +131,7 @@ pub(crate) fn build_block<B: StorageBackend>(
         let mut builder = BlockBuilder::new(
             Parents::new(parents).map_err(|e| ApiError::DependencyError(DependencyError::InvalidBlock(e)))?,
         )
-        .with_nonce_provider(nonce, 0f64);
+        .with_nonce_provider(nonce, 0);
         if let Some(payload) = payload {
             builder = builder.with_payload(payload)
         }
@@ -149,7 +149,7 @@ pub(crate) fn build_block<B: StorageBackend>(
         )
         .with_nonce_provider(
             MinerBuilder::new().with_num_workers(num_cpus::get()).finish(),
-            args.protocol_config.minimum_pow_score(),
+            args.protocol_config.minimum_pow_score() as u32,
         );
         if let Some(payload) = payload {
             builder = builder.with_payload(payload)
